Use bytemuck for no-padding struct byte views
