+++ crates/pointer_replacer/src/lib.rs
@@ -28,7 +28,7 @@ mod analyses;
 mod rewriter;
 mod utils;
 
-pub use rewriter::{Config, replace_local_borrows, rewrite_struct_arrays};
+pub use rewriter::{BytemuckDependency, Config, replace_local_borrows, rewrite_struct_arrays};
 
 #[cfg(test)]
 mod tests;
