+++ crates/utils/src/lib.rs
@@ -25,6 +25,7 @@ extern crate thin_vec;
 
 pub mod ast;
 pub mod bit_set;
+pub mod bytemuck;
 pub mod c_lib;
 pub mod compilation;
 pub mod disjoint_set;
